apps/api/pyproject.toml
[project]
name = "pulso-politico-api"
version = "0.1.5"
description = "API inicial do Pulso Politico"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "email-validator>=2.2.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "redis>=5.2.0",
  "pyjwt>=2.9.0",
  "alembic>=1.14.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.28.0"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
